+++ .coveragerc
@@ -8,7 +8,7 @@ omit =
   xonsh/lazyasd.py
   xonsh/parser_table.py
   xonsh/completion_parser_table.py
-  xonsh/ply/*
+  xonsh/parsers/ply/*
 
 # keep all cache files in one place
 data_file = .cache/coverage
