+++ xonsh/parsers/ply/lex.py
