+++ xonsh/parsers/v310.py
@@ -11,7 +11,7 @@ handle
 import ast
 
 from xonsh.parsers.v39 import Parser as ThreeNineParser
-from xonsh.ply.ply import yacc
+from xonsh.parsers.ply import yacc
 
 
 class Parser(ThreeNineParser):
